Raspberry/requirements.txt
Adafruit-DHT==1.3.4
python-decouple==3.1
paho-mqtt==1.4.0
PyMySQL==0.9.2
simplejson==3.16.0